add global support for goal start
